Fix std::format with protobuf enum in MessageDispatcher
std::format cannot format protobuf-generated enums (no formatter
specialization). Cast msg->messagetype() to int, matching the
handlers_ lookup on the previous line.

+++ common/MessageDispatcher.cpp
@@ -8,7 +8,7 @@ void MessageDispatcher::registerHandler(matrixserver::MessageType type, Handler 
 void MessageDispatcher::dispatch(std::shared_ptr<UniversalConnection> conn, std::shared_ptr<matrixserver::MatrixServerMessage> msg) {
     const auto it = handlers_.find(static_cast<int>(msg->messagetype()));
     if (it == handlers_.end()) {
-        BOOST_LOG_TRIVIAL(trace) << std::format("[MessageDispatcher] No handler for message type {} — dropping", msg->messagetype());
+        BOOST_LOG_TRIVIAL(trace) << std::format("[MessageDispatcher] No handler for message type {} — dropping", static_cast<int>(msg->messagetype()));
         return;
     }
     it->second(std::move(conn), std::move(msg));
